OpenCV changes
Not of this code is tested, and might not build properly. It should, though, be mainly sufficient once the values are fine-tuned to work well with the rings. At the moment, this also computes the contours, which can be useful for debugging purposes. It all logs the frames to the camera stream, after the contour finding procedure.