Updates from Match 3
Camera connection is working and seems to be correctly identifying ring stacks, however, it remains to be tested on the robot.
The TeleOp23 is all set, but after running our matches we'll make a few changes for driver ease (i.e. slow mode for tank drive). Also, strafing in little strokes is quite inaccurate: that is a mechanical issue, but I'll check whether we can use RoadRunner to stabilize that.
PIDAutonomous is working translationally but not rotationally. The issue with incorrect angles is most likely a product of having mounted the hub vertically (I thought I'd accounted for this but one of the axes must be off), so we need to run a few tests to determine what the new axes are so that RoadRunner propoerly adjusts heading instead of aimlessly searching around until it flips itself by either 90 or 180 degrees then progressing to the translational movement.