State Autonomous Code: Finished Wobble Deposit
Final Autonomous is the autonomous file we will be using for state.
Wobble deposit for all three positions is no longer time based: it is now fully Road Runner/encoder based.
Please make thoughtful changes and comment out code instead of deleting it.
Tasks remaining:
Robot placement for powershots needs to be fine-tuned.
OpenCV misclassifies stacks of 1 and 4 rings about 1/3 of the time.