State Control Award Autonomous Updates
Wobble and Flywheel are now voltage compensated!
Launch line is finished!
Path for 0 rings is finished! We should use spline in a similar manner for 1 and 4 rings too as it does a great job of avoiding the ring stack.
We're getting 2 rings in the high goal and one too far up, this is because there are both some mechanics and programming stuff to fine tune.
Also, the traj11 code that was worked on last weekend wasn't actually running because it was set to null when it was first declared. So the code that was running was what I had previously coded, hence it wasn't strafing to the side as intended.